new models and datasets

+++ src/evaluate.py
@@ -24,6 +24,10 @@ from tqdm import trange
 from dataset import Dataset
 from model   import Model
 
+import nltk
+nltk.download('punkt')
+
+
 # =  =  =  =  =  =  =  =  =  =  =  Logging Setup  =  =  =  =  =  =  =  =  =  =  =  =  = 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
@@ -33,6 +37,17 @@ logging.basicConfig(
 )
 # =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  =  = 
 
+def do_model_prediction(dataset, model, batch_size):
+
+    model_predictions = []
+    for i in trange(0, len(dataset.data_plain), batch_size, leave=False):
+        batch_inputs  = dataset.data_plain[i:i+batch_size]
+        batch_outputs = model.generate(batch_inputs)
+        model_predictions.extend(batch_outputs)
+
+    return model_predictions
+
+
 def main(
         dataset_name : str = "",
         model_name   : str = "",
@@ -40,6 +55,7 @@ def main(
         prompt_index : int = 0,
         eval_lang    : list = None,
         eval_mode    : str = "zero_shot",
+        overwrite    : bool = False,
 ):
     
     logger.info("Dataset name: {}".format(dataset_name))
@@ -48,20 +64,27 @@ def main(
     logger.info("Prompt index: {}".format(prompt_index))
     logger.info("Model Lang (applicable to cross-lingual datasets): {}".format(eval_lang))
     logger.info("Evaluation mode: {}".format(eval_mode))
+    logger.info("Overwrite: {}".format(overwrite))
     logger.info("")
 
     # Load dataset and model
     dataset = Dataset(dataset_name, prompt_index, support_langs=eval_lang, eval_mode=eval_mode)
-    model   = Model(model_name)
 
-    # Infer with model
-    model_predictions = do_model_prediction(dataset, model, batch_size)
-    data_with_model_predictions = dataset.dataset_processor.format_model_predictions(dataset.data_plain, model_predictions)
+    if overwrite or not os.path.exists('log/{}/{}/{}_p{}.json'.format(model_name, dataset_name, eval_mode, prompt_index)):
+        # Infer with model
+        model   = Model(model_name)
+        model_predictions = do_model_prediction(dataset, model, batch_size)
+        data_with_model_predictions = dataset.dataset_processor.format_model_predictions(dataset.data_plain, model_predictions)
 
-    # Save the result with predictions
-    os.makedirs('log/{}/{}'.format(eval_mode, dataset_name), exist_ok=True)
-    with open('log/{}/{}/{}_p{}.json'.format(eval_mode, dataset_name, "current_model", prompt_index), 'w') as f:
-        json.dump(data_with_model_predictions, f, indent=4, ensure_ascii=False)
+        # Save the result with predictions
+        os.makedirs('log/{}/{}'.format(model_name, dataset_name), exist_ok=True)
+        with open('log/{}/{}/{}_p{}.json'.format(model_name, dataset_name, eval_mode, prompt_index), 'w') as f:
+            try:
+                json.dump(data_with_model_predictions, f, indent=4, ensure_ascii=False)
+            except:
+                json.dump(data_with_model_predictions, f, indent=4, ensure_ascii=True)
+    
+    data_with_model_predictions = json.load(open('log/{}/{}/{}_p{}.json'.format(model_name, dataset_name, eval_mode, prompt_index), 'r'))
 
     # Metric evaluation
     results = dataset.dataset_processor.compute_score(data_with_model_predictions)
@@ -77,19 +100,11 @@ def main(
     print('\n\n\n')
 
     # Save the result with metrics
-    with open('log/{}/{}/{}_p{}_score.json'.format(eval_mode, dataset_name, "current_model", prompt_index), 'w') as f:
-        json.dump(results, f, indent=4, ensure_ascii=False)
-
-
-def do_model_prediction(dataset, model, batch_size):
-
-    model_predictions = []
-    for i in trange(0, len(dataset.data_plain), batch_size, leave=False):
-        batch_inputs  = dataset.data_plain[i:i+batch_size]
-        batch_outputs = model.generate(batch_inputs)
-        model_predictions.extend(batch_outputs)
-
-    return model_predictions
+    with open('log/{}/{}/{}_p{}_score.json'.format(model_name, dataset_name, eval_mode, prompt_index), 'w') as f:
+        try:
+            json.dump(results, f, indent=4, ensure_ascii=False)
+        except:
+            json.dump(results, f, indent=4, ensure_ascii=True)
 
 
 if __name__ == "__main__":
